fix tests: fix mixed up names and methods
