+++ Tests.PhraseTMS/ClientTests.cs
@@ -1,91 +1,34 @@
-﻿using PhraseTMSTests.Base;
-using Apps.PhraseTMS.Actions;
-using Apps.PhraseTMS.Constants;
-using Apps.PhraseTMS.Models.CustomFields;
-using Apps.PhraseTMS.Models.Jobs.Requests;
-using Apps.PhraseTMS.Models.Analysis.Requests;
-using Apps.PhraseTMS.Models.Projects.Requests;
+﻿using Apps.PhraseTMS.Actions;
+using Apps.PhraseTMS.Models.Clients.Requests;
 using Blackbird.Applications.Sdk.Common.Invocation;
+using PhraseTMSTests.Base;
 
 namespace Tests.PhraseTMS;
 
 [TestClass]
-public class AnalysisTests : TestBaseMultipleConnections
+public class ClientTests : TestBaseMultipleConnections
 {
-    public const string PROJECT_ID = "INLIOpS573UU4BbFBzs9v0";
-    public const string JOB_ID = "UkZvUPhvw1QItyADWZIPp3";
-    public const string ANALYSIS_ID = "JyaYpAIr8pF65xLsLfOOZ1";
+    public const string CLIENT_ID = "qdEcGUaxnLrs4d6z4B89c0";
 
-    [TestMethod, ContextDataSource(ConnectionTypes.Credentials)]
-    public async Task Create_analysis_works(InvocationContext context)
+    [TestMethod, ContextDataSource]
+    public async Task Search_clients_works(InvocationContext context)
     {
-        // Arrange
-        var actions = new AnalysisActions(context, FileManager);
-        var projectRequest = new ProjectRequest { ProjectUId = PROJECT_ID };
-        var analysis = new CreateAnalysisInput { JobsUIds = new List<string> { JOB_ID } };
-        var workflowStep = new WorkflowStepOptionalRequest { };
-        var jobsQuery = new ListAllJobsQuery { };
+        var actions = new ClientActions(context);
 
-        // Act
-        var result = await actions.CreateAnalysis(projectRequest, analysis, workflowStep, jobsQuery);
+        var result = await actions.ListClients(new ListClientsQuery { });
 
-        // Assert
         PrintResult(result);
-        Assert.IsTrue(result.Analyses.Any() && result.Analyses.All(x => x.Uid != null));
+        Assert.IsTrue(result.Clients.Any() && result.Clients.All(x => x.UId != null));
     }
 
     [TestMethod, ContextDataSource]
-    public async Task Search_project_analyses_works(InvocationContext context)
+    public async Task Get_client_works(InvocationContext context)
     {
-        // Arrange
-        var actions = new AnalysisActions(context, FileManager);
-        var projectRequest = new ProjectRequest { ProjectUId = PROJECT_ID };
+        var actions = new ClientActions(context);
 
-        // Act
-        var result = await actions.ListProjectAnalyses(projectRequest, new ListAnalysesQueryRequest { });
+        var result = await actions.GetClient(new ClientRequest { ClientUid = CLIENT_ID });
 
-        // Assert
         PrintResult(result);
-        Assert.IsTrue(result.Analyses.Any() && result.Analyses.All(x => x.Uid != null), "Project has no analyses");
-    }
-
-    [TestMethod, ContextDataSource]
-    public async Task Get_analysis_works(InvocationContext context)
-    {
-        // Arrange
-        var actions = new AnalysisActions(context, FileManager);
-
-        // Act
-        var result = await actions.GetJobAnalysis(new GetAnalysisRequest { AnalysisUId = ANALYSIS_ID });
-
-        // Assert
-        PrintResult(result);
-        Assert.IsTrue(result.Uid != null);
-    }
-
-    [TestMethod, ContextDataSource]
-    public async Task Download_analysis_file_works(InvocationContext context)
-    {
-        // Arrange
-        var actions = new AnalysisActions(context, FileManager);
-
-        // Act
-        var result = await actions.DownloadAnalysis(new GetAnalysisRequest { AnalysisUId = ANALYSIS_ID }, "JSON", null);
-
-        // Assert
-        PrintResult(result);
-    }
-
-    [TestMethod, ContextDataSource]
-    public async Task SetDateCustomField_works(InvocationContext context)
-    {
-        // Arrange
-        var actions = new CustomFieldsActions(context);
-        var date = DateTime.UtcNow.AddDays(15);
-        var project = new ProjectRequest { ProjectUId = "0SBo723Ge0wHfk0A1k1XWn0" };
-        var dateCustomField = new DateCustomFieldRequest { FieldUId = "gtCnCd6aZ0SkaGXu8wETa1" };
-
-        // Act
-        await actions.SetDateCustomField(project, dateCustomField, date);
+        Assert.IsNotNull(result);
     }
 }
